requirements.txt
pickles==0.1.1
PyYAML==6.0
Pillow==9.4.0

